% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sTabl3R.v.0.3.R
\name{check_input}
\alias{check_input}
\title{Check input for \code{\link[=generate_statistics]{generate_statistics()}} function.}
\usage{
check_input(df, group)
}
\arguments{
\item{df}{A \code{\link[=data.frame]{data.frame()}} object. See \code{FORMATTING} in \code{\link[=generate_statistics]{generate_statistics()}}.}

\item{group}{A grouping variable.}
}
\value{
A data frame, modified as necessary, to be used as input for the
\code{\link[=generate_statistics]{generate_statistics()}} function.
}
\description{
Checks to determine if the input is a data frame, whether the first column
contains unique IDs, and whether the grouping variable is present. If the
first row of the data frame does not contain unique IDs, these will be
supplied. If the grouping variable isn’t already a factor, it will be be
coerced to one.
}
\details{
Error checking is currently limited, and will be expanded as testing
continues.
}
\examples{
data(mtcars)
check_input(mtcars) # Generates error messages

# Make grouping variable
mtcars$Group <- rep("Group1", nrow(mtcars))
check_input(mtcars, group = "Group") # Warning message about identifiers
}
