% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sTabl3R.v.0.5.R
\name{extract_stats}
\alias{extract_stats}
\title{Extract and Display Statistics from a Dataframe}
\usage{
extract_stats(df, group = "Group", var)
}
\arguments{
\item{df}{A \code{\link[=data.frame]{data.frame()}} object. See \code{FORMATTING} in \code{\link[=generate_statistics]{generate_statistics()}}.}

\item{group}{A character string specifying the grouping variable. Must be a
column in the dataframe. Defaults to group = "Group".}

\item{var}{A character string specifying the variable for which to extract
statistics. Must be a column in the dataframe.}
}
\value{
NULL. The function prints the statistics to the console.
}
\description{
This function tries to extract statistics from a dataframe using
\code{sTabl3R} functions based on a specified grouping variable and a second variable
in the dataframe. It then displays the statistics in the console in a
user-friendly format.
}
\examples{
data <- data.frame(id = c("Observation 1","Observation 2",
        "Observation 3","Observation 4", "Observation 5","Observation 6",
        "Observation 7","Observation 8"),
        Group = c("A", "B", "A", "B", "A", "B", "A", "B"),
        Var = c(1, 2, 3, 4, 5, 6, 7, 8))
extract_stats(data, "Group", "Var")

data(mtcars)
extract_stats(mtcars, group = "cyl", var = "disp")
}
