% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sTabl3R.v.0.2.R
\name{generate_results_tables}
\alias{generate_results_tables}
\title{Generate results tables from \code{\link[=generate_statistics]{generate_statistics()}} output.}
\usage{
generate_results_tables(results)
}
\arguments{
\item{results}{A list from \code{\link[=generate_statistics]{generate_statistics()}} that contains the
results of the statistical tests. It should be an S3 object of class \code{sTable}
for multiple group comparisons or \code{ssTable} for single group comparisons.}
}
\value{
A list of flextables for each variable in the results. If the input
results are of class 'sTable', it returns a list of flextables for the
continuous and/or categorical variables. If the input results are of class
'ssTable', it returns a flextable for the continuous variables and a list
of flextables for each categorical variable.
}
\description{
This function generates tables using \code{knitr::kable} and \code{flextable} for
the results of statistical tests. It supports both single group and
multiple group comparisons. For single group comparisons, it generates descriptive summary statistics.
For multiple group comparisons, it generates summary tables with group
comparison statistics.
}
\examples{
results <- generate_statistics(data, group = "My_Grouping_Variable")
tables <- generate_results_tables(results)
}
