% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sTabl3R.v.0.6.0.0.R
\name{summary.sTable}
\alias{summary.sTable}
\title{Summarize an \code{sTable} object}
\usage{
\method{summary}{sTable}(x, ...)
}
\arguments{
\item{x}{An \code{sTable} object}

\item{...}{Additional arguments}
}
\value{
No return value; this function is called for printing to the console
}
\description{
This function provides a summary of an \code{sTable} object, including the groups,
the variables that were categorized as continuous (if present), and the variables
that were deemed to be categorical (if present) when the \code{\link[=generate_statistics]{generate_statistics()}}
function was called.
}
