% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sTabl3R.v.0.5.R
\name{generate_statistics}
\alias{generate_statistics}
\title{Generate statistical summary of a data frame.}
\usage{
generate_statistics(df, group = "Group", force_nonparametric = F)
}
\arguments{
\item{df}{A \code{\link[=data.frame]{data.frame()}} object.}

\item{group}{A character string specifying the grouping variable. Must be a column in the dataframe.
Defaults to group = "Group".}

\item{force_nonparametric}{Logical. If TRUE, the function will use non-parametric statistical testing. Default is FALSE.}
}
\value{
The function returns a list containing counts, continuous variable tables
with associated statistical results, and, for categorical variables,
contingency tables with associated statistical results. The formatting of
the output differs slightly depending on the number of levels in the grouping
variable.

The output is an informal S3 list of class \code{sTable} or \code{ssTable} that may
either be examined directly or used as input for the \code{\link[=generate_results_tables]{generate_results_tables()}}
function.

Structure of the \code{sTable} class:

List of 3      :
$ Counts      :   data.frame of counts (samples per group)
$ Continuous  :   if present, summary tables and statistics for each variable
$ Categorical :   if present, contingency tables and statistics for each variable

Structure of the \code{ssTable} class:

List of 3      :
$ Counts      :   data.frame of counts (number of samples)
$ Continuous  :   if present, summary table of continuous variables
$ Categorical :   if present, single group contingency tables for each variable
}
\description{
This function will perform a quick statistical analysis on a supplied
data frame (df) using a specified grouping variable (group). Overall, this
function is intended to automate the process of simple statistical analysis
for a data set, and may be of use in exploratory data analysis.
}
\details{
Following limited error checking, the function attempts to identify
numeric (continuous and ordinal) and categorical variables within the
data frame.

For each identified numeric variable, Shapiro-Wilk testing is employed to
check for normality. Then, each numeric variable is subjected either to
parametric (t-test or one-way ANOVA) or non-parametric (Mann-Whitney U test
or Kruskal-Wallis test) comparisons depending on the number of
strata in the grouping variable. For instances where the grouping variable
has a single level, only summary statistics will be generated.

For categorical variables, the function will apply either the Chi-squared test
or the Fisher’s Exact test based on the expected frequencies in each contingency
table, unless the grouping variable has a single level, in which case only
summary tables will be generated.

\code{FORMATTING}
The data frame should be an R \code{\link[=data.frame]{data.frame()}} object in which the first
column is a list of unique identifiers. The \code{group} argument should identify
a column in the data frame that will be used to distinguish the groups that
are to be compared statistically. A single group can be generated as shown in
the examples below for the \code{mtcars} dataset.

}
\examples{
data(mtcars)
# Generate a grouping variable
mtcars$Group <- as.factor(rep("GroupA", nrow(mtcars))) # Single group case
res1 <- generate_statistics(mtcars, group = "Group")
summary(res1)
generate_results_tables(res1)

data(mtcars)
res2 <- generate_statistics(mtcars, group = "cyl") # Multiple group levels
summary(res2) 
generate_results_tables(res2)
}
