% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sTabl3R.v.0.4.R
\name{summary.ssTable}
\alias{summary.ssTable}
\title{Summarize an \code{ssTable}object}
\usage{
\method{summary}{ssTable}(x, ...)
}
\arguments{
\item{x}{An sTable object}

\item{...}{Additional arguments}
}
\value{
No return value; this function is called for printing to the console
}
\description{
This function provides a summary of an \code{ssTable} object, including the group,
the variables that were deemed to be continuous (of present), and the variables
that were deemed to be categorical (if present) when the \code{\link[=generate_statistics]{generate_statistics()}}
function was called.
}
